requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.0
frozenlist==1.3.1
h11==0.14.0
idna==3.4
motor==3.0.0
multidict==6.0.2
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
pytz==2022.5
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
yarl==1.8.1

celery==5.2.7

pandas==1.5.2
matplotlib==3.6.2
kneed==0.7.0
scipy==1.9.3
numpy==1.24.0
sklearn==0.0

#gevent==22.10.2

requests==2.28.1
scikit-learn==1.2.0
#asgiref==3.6.0
kombu==5.2.4